# rms-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RMS API Documentation",
        "description": "API du Revenu Minimum Solidaire (RMS) : auth, portefeuille, file FIFO, rotations, timeline.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/timeline/posts": {
            "get": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Lister les posts de la timeline (feed paginé)",
                "operationId": "ec21f6d06620cdbd195fd0ba084d380f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de post",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des posts"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Créer un post dans la timeline",
                "operationId": "d9162c35234ef034f6cc044ba201f025",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Grâce au RMS, j'ai lancé mon petit commerce."
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "text"
                                    },
                                    "media_url": {
                                        "type": "string",
                                        "example": "https://example.com/photo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post créé"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/timeline/posts/{post}/like": {
            "post": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Liker ou déliker un post",
                "operationId": "87da051eae463a27f460c6dac44a535f",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like togglé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/timeline/posts/{post}/comment": {
            "post": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Ajouter un commentaire sur un post",
                "operationId": "636e222c2c474c14ac9f50658f92a315",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Bravo pour ta réussite !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}